Check correctly
