+++ daft/execution/udf_worker.py
@@ -1,11 +1,11 @@
 from __future__ import annotations
 
-import pickle
 import sys
 from multiprocessing.connection import Client
 from pickle import PicklingError
 from traceback import TracebackException
 
+import daft.pickle
 from daft.errors import UDFException
 from daft.execution.udf import (
     _ENTER,
@@ -45,7 +45,7 @@ def udf_event_loop(
 
             # We initialize after ready to avoid blocking the main thread
             if expression_projection is None:
-                uninitialized_projection: ExpressionsProjection = pickle.loads(expr_projection_bytes)
+                uninitialized_projection: ExpressionsProjection = daft.pickle.loads(expr_projection_bytes)
                 initialized_projection = ExpressionsProjection([e._initialize_udfs() for e in uninitialized_projection])
                 expression_projection = initialized_projection
 
@@ -66,7 +66,7 @@ def udf_event_loop(
         try:
             # TODO: Consider using cloudpickle, since it can pickle more types
             # like lambda functions
-            exc_bytes = pickle.dumps(exc)
+            exc_bytes = daft.pickle.dumps(exc)
         except (PicklingError, AttributeError):
             exc_bytes = None
         conn.send((_UDF_ERROR, e.message, TracebackException.from_exception(exc), exc_bytes))
