+++ daft/pickle/pickle.py
@@ -1,12 +1,14 @@
 from __future__ import annotations
 
-from daft.pickle.cloudpickle_fast import dumps as cloudpickle_dumps
-from daft.pickle.cloudpickle_fast import loads as cloudpickle_loads
+from typing import Any
+
+from daft.pickle.cloudpickle import dumps as cloudpickle_dumps  # type: ignore
+from daft.pickle.cloudpickle import loads as cloudpickle_loads  # type: ignore
 
 
-def dumps(obj: object) -> bytes:
+def dumps(obj: Any) -> bytes:
     return cloudpickle_dumps(obj)
 
 
-def loads(data: bytes) -> object:
+def loads(data: bytes) -> Any:
     return cloudpickle_loads(data)
